eaclib + game GTE: 6-agent parallel matching run (+18 byte-matches)
Sonnet-5 fan-out (1 agent per eaclib module + 3 GTE clusters), Opus-consolidated:
every claimed PASS re-verified via tools/verify_asm.py, whole batch compile-clean.

  eacpsxz  (2): strrstr; CD_Stopread [+ CD-context struct materialized]
  sndpsxz  (7): iSNDserveradd100hzclient, iSNDserveraddclient, iSNDbankalloc,
                cSNDplay, SNDPKTPLAY_stop, iSNDstreamparsenumchunks, iSNDstreamgetrequestptr
  spchpsxz (5): iSPCH_SearchEventDat, iSPCH_SaveChosenSentence, iSPCH_ClearCycleBit,
                SPCH_GetSampleDataRate, iSPCH_GetRuleID
  weather  (4): Weather_ChangeDensityState, Weather_ChangeIntensityState,
                Weather_ChangeDensityBasedOnTime, Weather_InitStateControls

Plus many near-miss reductions (draw/draww/flare/hrzsku/sfx/trsproj + sndpsxz/spchpsxz
floors) and real bug fixes: DrawW_kCtrlWorld_High quad-walk, iSNDunsafevol return-0,
iSPCH_BankMemAlloc return-value + 2-arg sig, dead-local cleanups.

tools/verify_asm.py: (1) resolve co-equal aliases name->addr->disasm-label; (2) SKIP
*UND*/*ABS* symbols so an undefined extern decl at addr 0 no longer false-resolves to
the function objdump labels at offset 0.

+++ recon/eaclib/psx/spchpsxz/spchevnt.cpp
@@ -68,7 +68,7 @@ extern "C" unsigned int VoxEvent_GetKeepTillExpiresFlag(int e)
 /* iSPCH_SearchEventDat @0x800E6EC4 : address of the entry in blob `dat` whose id == eventID, or 0. */
 extern "C" int iSPCH_SearchEventDat(int dat, unsigned int eventID)
 {
-    unsigned short count = *(unsigned short *)(dat + 2);
+    unsigned int count = *(unsigned short *)(dat + 2);
     int table = 0;
     if (count != 0) {
         do {
@@ -76,7 +76,7 @@ extern "C" int iSPCH_SearchEventDat(int dat, unsigned int eventID)
             table = table + 1;
             if (*p == eventID)
                 return (int)p;
-        } while (table < (int)(unsigned int)count);
+        } while (table < (int)count);
     }
     return 0;
 }
